Added error checks for webscraping, improved at_bats.get_player_abs function, added scraping tests. All tests are currently passing.
Added error checking in bio.py and at_bats.py to ensure the webdriver is properly working and, if not, to create a new webdriver.
Added error checking in at_bats.py to test if the URL can be acquired. If not, retries up to 5 times, then errors out.
Altered the get_player_abs function to accept multiple players and return them as a list of dictionaries.
Altered get_player_abs() to allow more flexibility to the storage of existing at bat data and player bio data.
Added tests to test_at_bats.py to test scraping with and without an existing webdriver and to test the errors that could be encountered.