Added test for saving data in CSVs. Made some fixes to at_bat.py as a result.
Added test_save_at_bats() to test_at_bats.py. Tests that data is a properly saved.
Made change to save data in at_bats.py if the data is found in the existing data filepath provided. Old: didn't even try to save this data. New: Users may want to save the data to a new location, so if a save folder is provided, it will save the existing data again.